Fix codepoint pages
