+++ src/routes/codepoint/[id].json/+server.ts
@@ -2,14 +2,15 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import type { Request, Response, NextFunction } from "express";
 import { lookupChar } from "$lib/server/Unicode";
+import { error, json, type RequestHandler } from "@sveltejs/kit";
 
-export function get(req: Request, res: Response, next: NextFunction) {
-	const { id } = req.params;
+export const GET = (async ({ url, params }) => {
+	const { id } = params;
 	const char = lookupChar(parseInt(id, 16));
 	if (!char) {
-		return next();
+		throw error(404);
 	}
-	res.json(char);
-}
+
+	return json(char);
+}) satisfies RequestHandler;
